+++ .storybook/main.ts
@@ -7,6 +7,8 @@ const config: StorybookConfig = {
 		'@storybook/addon-essentials',
 		'@storybook/addon-interactions',
 		'@storybook/addon-styling',
+		'@storybook/addon-a11y',
+		// '@chromaui/addon-visual-tests',
 	],
 	framework: '@storybook/sveltekit',
 	docs: {