+++ scripts/pagefind.js
@@ -11,15 +11,9 @@ await index.addDirectory({
 });
 
 await index.writeFiles({
-	outputPath: 'src/pagefind',
-});
-
-await index.writeFiles({
 	outputPath: 'static/pagefind',
 });
 
-// Write ts-nocheck to the top of the file
-import fs from 'fs';
-const file = 'src/pagefind/pagefind.js';
-const data = fs.readFileSync(file);
-fs.writeFileSync(file, '// @ts-nocheck\n' + data);
+await index.writeFiles({
+	outputPath: 'build/pagefind',
+});